# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth Login Protect API",
    "version": "1.0.0",
    "description": "Authentication API backed by Supabase, with public and JWT-protected routes."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully." },
          "400": { "description": "Email and password are required, or signup failed." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive access and refresh tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful, returns access_token and refresh_token." },
          "400": { "description": "Email and password are required." },
          "401": { "description": "Invalid login credentials." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Exchange a refresh token for a new access token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": { "type": "string" }
                },
                "required": ["refresh_token"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Refresh successful, returns a new access_token and refresh_token." },
          "400": { "description": "Refresh token is required." },
          "401": { "description": "Invalid or expired refresh token." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "Logout successful, no content." },
          "400": { "description": "Logout failed." },
          "401": { "description": "Missing, invalid, or expired access token." },
          "500": { "description": "Internal server error." }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public information endpoint",
        "responses": {
          "200": { "description": "Successfully retrieved public info." }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the authenticated user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Successfully retrieved user profile." },
          "401": { "description": "Missing, invalid, or expired access token." }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get the authenticated user's dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Successfully retrieved dashboard." },
          "401": { "description": "Missing, invalid, or expired access token." }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": { "description": "Service is healthy." }
        }
      }
    }
  }
}